Extract model metadata into codex-models
