+++ codex-rs/core/src/auth_env_telemetry.rs
@@ -1,9 +1,10 @@
+use codex_models::ModelProviderInfo;
+use codex_models::WireApi;
 use codex_otel::AuthEnvTelemetryMetadata;
 
 use crate::auth::CODEX_API_KEY_ENV_VAR;
 use crate::auth::OPENAI_API_KEY_ENV_VAR;
 use crate::auth::REFRESH_TOKEN_URL_OVERRIDE_ENV_VAR;
-use crate::model_provider_info::ModelProviderInfo;
 
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
 pub(crate) struct AuthEnvTelemetry {
@@ -64,7 +65,7 @@ mod tests {
             env_key: Some("sk-should-not-leak".to_string()),
             env_key_instructions: None,
             experimental_bearer_token: None,
-            wire_api: crate::model_provider_info::WireApi::Responses,
+            wire_api: WireApi::Responses,
             query_params: None,
             http_headers: None,
             env_http_headers: None,
