+++ codex-rs/core/src/models_manager/mod.rs
@@ -1,15 +1,16 @@
-pub mod cache;
-pub mod collaboration_mode_presets;
 pub mod manager;
 pub mod model_info;
-pub mod model_presets;
+pub mod cache {
+    pub use codex_models::models_manager::cache::*;
+}
+pub mod collaboration_mode_presets {
+    pub use codex_models::models_manager::collaboration_mode_presets::*;
+}
+pub mod model_presets {
+    pub use codex_models::models_manager::model_presets::*;
+}
 
 /// Convert the client version string to a whole version string (e.g. "1.2.3-alpha.4" -> "1.2.3").
 pub fn client_version_to_whole() -> String {
-    format!(
-        "{}.{}.{}",
-        env!("CARGO_PKG_VERSION_MAJOR"),
-        env!("CARGO_PKG_VERSION_MINOR"),
-        env!("CARGO_PKG_VERSION_PATCH")
-    )
+    codex_models::models_manager::client_version_to_whole()
 }
