+++ codex-rs/models/BUILD.bazel
@@ -1,0 +1,9 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "models",
+    crate_name = "codex_models",
+    compile_data = glob(["templates/**"]) + [
+        "//codex-rs/core:models.json",
+    ],
+)
