+++ codex-rs/models/src/lib.rs
@@ -1,0 +1,18 @@
+pub mod model_provider_info;
+pub mod models_manager;
+pub mod response_debug_context;
+
+pub use model_provider_info::DEFAULT_LMSTUDIO_PORT;
+pub use model_provider_info::DEFAULT_OLLAMA_PORT;
+pub use model_provider_info::DEFAULT_WEBSOCKET_CONNECT_TIMEOUT_MS;
+pub use model_provider_info::EnvVarError;
+pub use model_provider_info::LEGACY_OLLAMA_CHAT_PROVIDER_ID;
+pub use model_provider_info::LMSTUDIO_OSS_PROVIDER_ID;
+pub use model_provider_info::ModelProviderInfo;
+pub use model_provider_info::OLLAMA_CHAT_PROVIDER_REMOVED_ERROR;
+pub use model_provider_info::OLLAMA_OSS_PROVIDER_ID;
+pub use model_provider_info::OPENAI_PROVIDER_ID;
+pub use model_provider_info::WireApi;
+pub use model_provider_info::built_in_model_providers;
+pub use model_provider_info::create_oss_provider;
+pub use model_provider_info::create_oss_provider_with_base_url;
