+++ codex-rs/models/src/model_provider_info_tests.rs
