+++ codex-rs/models/src/models_manager/cache.rs
@@ -11,27 +11,24 @@ use tokio::fs;
 use tracing::error;
 use tracing::info;
 
-/// Manages loading and saving of models cache to disk.
 #[derive(Debug)]
-pub(crate) struct ModelsCacheManager {
+pub struct ModelsCacheManager {
     cache_path: PathBuf,
     cache_ttl: Duration,
 }
 
 impl ModelsCacheManager {
-    /// Create a new cache manager with the given path and TTL.
-    pub(crate) fn new(cache_path: PathBuf, cache_ttl: Duration) -> Self {
+    pub fn new(cache_path: PathBuf, cache_ttl: Duration) -> Self {
         Self {
             cache_path,
             cache_ttl,
         }
     }
 
-    /// Attempt to load a fresh cache entry. Returns `None` if the cache doesn't exist or is stale.
-    pub(crate) async fn load_fresh(&self, expected_version: &str) -> Option<ModelsCache> {
+    pub async fn load_fresh(&self, expected_version: &str) -> Option<ModelsCache> {
         info!(
-                cache_path = %self.cache_path.display(),
-                expected_version,
+            cache_path = %self.cache_path.display(),
+            expected_version,
             "models cache: attempting load_fresh"
         );
         let cache = match self.load().await {
@@ -73,8 +70,7 @@ impl ModelsCacheManager {
         Some(cache)
     }
 
-    /// Persist the cache to disk, creating parent directories as needed.
-    pub(crate) async fn persist_cache(
+    pub async fn persist_cache(
         &self,
         models: &[ModelInfo],
         etag: Option<String>,
@@ -91,8 +87,7 @@ impl ModelsCacheManager {
         }
     }
 
-    /// Renew the cache TTL by updating the fetched_at timestamp to now.
-    pub(crate) async fn renew_cache_ttl(&self) -> io::Result<()> {
+    pub async fn renew_cache_ttl(&self) -> io::Result<()> {
         let mut cache = match self.load().await? {
             Some(cache) => cache,
             None => return Err(io::Error::new(ErrorKind::NotFound, "cache not found")),
@@ -123,14 +118,12 @@ impl ModelsCacheManager {
     }
 
     #[cfg(test)]
-    /// Set the cache TTL.
-    pub(crate) fn set_ttl(&mut self, ttl: Duration) {
+    pub fn set_ttl(&mut self, ttl: Duration) {
         self.cache_ttl = ttl;
     }
 
     #[cfg(test)]
-    /// Manipulate cache file for testing. Allows setting a custom fetched_at timestamp.
-    pub(crate) async fn manipulate_cache_for_test<F>(&self, f: F) -> io::Result<()>
+    pub async fn manipulate_cache_for_test<F>(&self, f: F) -> io::Result<()>
     where
         F: FnOnce(&mut DateTime<Utc>),
     {
@@ -143,8 +136,7 @@ impl ModelsCacheManager {
     }
 
     #[cfg(test)]
-    /// Mutate the full cache contents for testing.
-    pub(crate) async fn mutate_cache_for_test<F>(&self, f: F) -> io::Result<()>
+    pub async fn mutate_cache_for_test<F>(&self, f: F) -> io::Result<()>
     where
         F: FnOnce(&mut ModelsCache),
     {
@@ -157,19 +149,17 @@ impl ModelsCacheManager {
     }
 }
 
-/// Serialized snapshot of models and metadata cached on disk.
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub(crate) struct ModelsCache {
-    pub(crate) fetched_at: DateTime<Utc>,
+pub struct ModelsCache {
+    pub fetched_at: DateTime<Utc>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) etag: Option<String>,
+    pub etag: Option<String>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub(crate) client_version: Option<String>,
-    pub(crate) models: Vec<ModelInfo>,
+    pub client_version: Option<String>,
+    pub models: Vec<ModelInfo>,
 }
 
 impl ModelsCache {
-    /// Returns `true` when the cache entry has not exceeded the configured TTL.
     fn is_fresh(&self, ttl: Duration) -> bool {
         if ttl.is_zero() {
             return false;
