+++ codex-rs/models/src/models_manager/collaboration_mode_presets_tests.rs
