+++ codex-rs/models/src/models_manager/model_presets.rs
